fix: developer assets
